updated doom to x64

+++ inc/doom.h
@@ -340,7 +340,7 @@ typedef struct s_time
 	float			delta;
 	SDL_Color		color;
 	SDL_Surface		*surf;
-	long			start;
+	long long int   start;
 	struct tm		date;
 	SDL_Color		clock_fg;
 	SDL_Color		clock_bg;
